// github/types/components/schemas/organization-simple.ts
/**
 * A GitHub organization.
 */
export type organization_simple = {
  /**
   * @example "github"
   */
  login: string;
  /**
   * @example 1
   */
  id: number;
  /**
   * @example "MDEyOk9yZ2FuaXphdGlvbjE="
   */
  node_id: string;
  /**
   * @format uri
   * @example "https://api.github.com/orgs/github"
   */
  url: string;
  /**
   * @format uri
   * @example "https://api.github.com/orgs/github/repos"
   */
  repos_url: string;
  /**
   * @format uri
   * @example "https://api.github.com/orgs/github/events"
   */
  events_url: string;
  /**
   * @example "https://api.github.com/orgs/github/hooks"
   */
  hooks_url: string;
  /**
   * @example "https://api.github.com/orgs/github/issues"
   */
  issues_url: string;
  /**
   * @example "https://api.github.com/orgs/github/members{/member}"
   */
  members_url: string;
  /**
   * @example "https://api.github.com/orgs/github/public_members{/member}"
   */
  public_members_url: string;
  /**
   * @example "https://github.com/images/error/octocat_happy.gif"
   */
  avatar_url: string;
  /**
   * @example "A great organization"
   */
  description: string;
};